fix homepage animations
